refactor: 폴더 구조 변경
- 기존 client에서만 활용되던 폴더들을 client 폴더로 이동
- src 하위에는 client와 server로 나누어 관리
