+++ src/app/api/auth/naver/callback/route.ts
@@ -1,6 +1,6 @@
-import { getAccessToken, getUserInfo } from '@/src/utils/auth';
-import { createJwtToken } from '@/src/utils/jwt';
-import { ERROR_INFOS } from '@/src/constants/ERROR_INFOS';
+import { getAccessToken, getUserInfo } from '@/src/client/utils/auth';
+import { createJwtToken } from '@/src/client/utils/jwt';
+import { ERROR_INFOS } from '@/src/client/constants/ERROR_INFOS';
 import {
   createErrorApiResponse,
   createSuccessApiResponse,
