+++ src/app/products/[id]/components/FloatingCard.tsx
@@ -1,7 +1,7 @@
 import { Box, Button, Flex, Text } from '@chakra-ui/react';
 import VerifiedIconSVG from '@/public/assets/verify_badge.svg';
 import Image from 'next/image';
-import { formatKoreanCurrency } from '@/src/utils/currency';
+import { formatKoreanCurrency } from '@/src/client/utils/currency';
 
 interface FloatingCardProps {
   expertName: string;
