+++ src/client/config/auth.ts
