+++ src/client/errors/AppError.ts
