+++ src/client/types/accordionItem.ts
